refactor(server): move engine unit tests into engine/tests.rs
Extract the cfg(test) module so production engine.rs stays near the 1k-line boundary after party polish landed on main.
